% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broennimannOverlap.R
\name{broennimannEcospat}
\alias{broennimannEcospat}
\title{broennimannEcospat}
\usage{
broennimannEcospat(native_stack, invasive_stack, native_occ, invasive_occ,
  kept_axes = 2)
}
\arguments{
\item{native_stack}{A raster stack of the environmental data for the 
range of the native species.}

\item{invasive_stack}{A raster stack of the environmental data for the
range being invaded}

\item{kept_axes}{The nf parameter in the function dudi.pca - is the number
of kept axes - defaults to 2.}

\item{natice_occ}{Occurence points for the native species (i.e. where the 
species that will be invading into the invasive range is in it's native
range). A two-column matrix with column headings x and y (corresponding to
longitude and lattitude)}

\item{invasice_occ}{Occurence points for the species in the invasive range
(either the invading species in it's invasive range, OR the species that 
might compete with the invader).A two-column matrix with column headings 
x and y (corresponding to longitude and lattitude)}
}
\description{
Takes raster layers and occurence data and fits the broennimann
niche similarity, equivalency and overlap models.
}
