% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbnHarvest.R
\name{findDatasets}
\alias{findDatasets}
\title{findDatasets}
\usage{
findDatasets(nums)
}
\arguments{
\item{nums}{A vector of numbers to test.}
}
\description{
A shortcut function to apply isDataset over a large range of potential
dataset numbers - returns a 2D array containing the numbers of actual
datasets (i.e. excluding species lists and things that aren't anything).
}
