% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbnHarvest.R
\name{getAllData}
\alias{getAllData}
\title{getAllData}
\usage{
getAllData(numrange, filename, reference_list = NULL, links = NULL,
  links_only = FALSE, ...)
}
\arguments{
\item{numrange}{A range of numbers to test to see if they are datasets, and
get data from.}

\item{filename}{The name of the file that the data is going to be written
into - prefixes _links.csv for the links file.}

\item{reference_list}{A list of species to get data for.}

\item{links}{If specified then data is downloaded from these links.}

\item{links_only}{If TRUE then no data is downloaded, but the links are
gathered.}
}
\description{
Wraps all the above functions together - give it a range of numbers and it
will check if they are datasets, get the links then download the data.
Can set to "links only" mode, which will download the links, but not get the
data. To go hand in hand with this, it can also be supplied with some
links, in which case it won't bother testing to see if a dataset is a dataset
and so on...
}
