% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbnHarvest.R
\name{getLinks}
\alias{getLinks}
\title{getLinks}
\usage{
getLinks(num, reference_list, print = FALSE, filename)
}
\arguments{
\item{num}{A valid dataset number.}

\item{reference_list}{A list of species names that links should be collected
for.}

\item{print}{If TRUE then the dataset number if printed to screen - this is
mostly just for testing...}

\item{filename}{This is the prefix of the filename that the saved links are
written to.}
}
\description{
Collects links to individual records from valid dataset IDs. It does this
by collecting the number of records, breaking that up into chunks of 1000,
and then going to the URL that displays links to individual records for the
first 1-1000 records, 1001-2000 records etc. and then saves those links,
writing them to the working directory. It writes them in chunks of 1000. I
write these links to the disk since a) it can be a super slow process and I
don't want to repeat it if something goes wrong, and b) I don't want to
repeat it over and over and over-query the servers at NBN. If a reference
list of species names is provided then only species records that match that
list will be saved.
}
