% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbnHarvest.R
\name{getLinksAllRecords}
\alias{getLinksAllRecords}
\title{getLinksAllRecords
I just noticed that the link https://records.nbnatlas.org/occurrence/search
seems to display all the records in the NBN - it seems like this might just
be a better way to go about this - move through all the records collecting
links for the species on the C4 list, and then download them. Since this
misses out the step where I find out the dataset information I think that
I need to make sure that I include this when downloading the actual data,
but that's another function to go hand-in hand with this one.}
\usage{
getLinksAllRecords(range, filename, max)
}
\description{
getLinksAllRecords
I just noticed that the link https://records.nbnatlas.org/occurrence/search
seems to display all the records in the NBN - it seems like this might just
be a better way to go about this - move through all the records collecting
links for the species on the C4 list, and then download them. Since this
misses out the step where I find out the dataset information I think that
I need to make sure that I include this when downloading the actual data,
but that's another function to go hand-in hand with this one.
}
