% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbnHarvest.R
\name{isDataset}
\alias{isDataset}
\title{isDataset}
\usage{
isDataset(num)
}
\arguments{
\item{num}{The number to test.}
}
\description{
Takes a number, and probes the URL with that number as the extension to see
if that number leads to a dataset, or not. This basically just works off the
fact that NBN records are all indexed at URLs with a varying number at the
end.
}
